feat: add new args for multiple jobs and move helper methods to utils
